+++ ORStools/proc/directions_points_layer_proc.py
@@ -132,6 +132,30 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
             QgsProcessingParameterBoolean(self.EXPORT_ORDER, self.tr("Export order of jobs")),
         ]
 
+        self.setToolTip(self.PARAMETERS[0], self.tr("Point or MultiPoint layer."))
+        self.setToolTip(
+            self.PARAMETERS[1],
+            self.tr(
+                "Values will transfer to the output layer and can be used to join layers or group features afterwards."
+            ),
+        )
+        self.setToolTip(
+            self.PARAMETERS[2],
+            self.tr(
+                "Before running the algorithm points are sorted by the values of this field (Be aware of the field type! Text fields will be sorted like 1,13,2,D,a,x)."
+            ),
+        )
+        self.setToolTip(
+            self.PARAMETERS[3],
+            self.tr("Dictates the cost. For longer routes don't use Shortest Path."),
+        )
+        self.setToolTip(
+            self.PARAMETERS[4],
+            self.tr(
+                "You can optionally perform a Traveling Salesman Optimization on the waypoints of each (Multi)Point feature. Enabling Traveling Salesman will erase all other advanced configuration and assume the preference to be fastest."
+            ),
+        )
+
     def processAlgorithm(
         self, parameters: dict, context: QgsProcessingContext, feedback: QgsProcessingFeedback
     ) -> Dict[str, str]: