+++ CHANGELOG.md
@@ -45,6 +45,7 @@ RELEASING:
 ### Added
 - pre-commit configuration with Ruff linter for code quality enforcement
 - Improve cursor behaviour during digitization
+- tooltip hints in processing algorithms ([#196](https://github.com/GIScience/orstools-qgis-plugin/issues/196))
 - Improve isochrone color ramp with many ranges
 
 ### Fixed